moved/renamed haloexchange.hpp
